JavaScript bindings: removed them.
Although it would be nice to have JavaScript "bindings" for libOpenCOR, those are not actual bindings since libOpenCOR needs to be built using Emscripten. This means that any third-party library used by libOpenCOR will have to be built using Emscripten. It should be fine for a library like SUNDIALS, but it's likely to be much more complicated for a library like LLVM-Clang. So, for now, we are leaving JavaScript "bindings" out.