Python bindings.
Don't get our Python bindings to link to our libOpenCOR library anymore. Indeed, although we were able to get it to work without any problems on macOS, it required a static library on Linux and, for some reasons, just didn't work on Windows. So, instead, we add all the sources/headers in our Python module, and voila!