Dropped the idea of file manager.
This is what OpenCOR does, but that's because we have to deal with a GUI. When it comes to libOpenCOR, we should allow for several instances of a `File` object referring to the same local/remote file. A user might indeed want a `File` object for a given local file, and another `File` object after s/he has modified the local file. Either way, we shouldn't restrict the use of `File` objects that point to the same local/remote file.