CMake: build as much as we can and in debug mode.
By default, we build shared libraries which means that we can't build our Python bindings. Also, we build in debug mode since this is what we need and want most of the time. Finally, we don't do code coverage and memory checks. The former because it's not compatible with building Python bindings (it requires shared libraries while our Python bindings require a static library) and the latter for consistency with code coverage.

Testing...

Revert "Testing..."

This reverts commit c3dbd5425b0820b7d771e581c382c7a712486120.

Testing...

Testing...