Tests: reverted the way we retrieve and release some memory objects.
We had previously changed things to address an issue with some memory blocks being altered (after calling File::contents() from different tests!?). However, our "fix" only worked temporarily. So, we clearly need to find the proper cause of the problem and, in the meantime, I prefer to revert to the way we used to do things since it's better in my view.