CI: get buildcache to use strict accuracy and no direct mode.
We used to use the default accuracy and enable the direct mode for buildcache. But, when reorganising our files (without modifying them), llvm-cov would think that the files were still in their old location. That's because it indirectly relies on buildcache for path information.

So, now, we use the strict accuracy and disable direct mode, and all is fine.