Solver: moved the solverforwardeuler.h back to the API.
This seems to be the only way to get our JavaScript to work when creating an unknown solver since we always get a Solver object no matter what. At least, now, we get either a SolverForwardEuler object (i.e. we have a Forward Euler solver) or a Solver object (i.e. we have an unknown solver).