API: removed SolverForwardEuler and SolverUnknown.
No need for them to be part of the API. They were initially put there because of the way the solvers were being created, and we needed to know the name of the class in our JavaScript bindings. However, that approach wasn't working with our Python bindings, so had to use another approach which, in the end, is cleaner since we don't need those classes in our API anymore.