Solvers: a solver is not to be used directly by an end-user of libOpenCOR.
It's like our Compiler class which is only meant to be used internally.

Also, to make our solvers usable by an end-user would be somewhat tricky since we need to provide the computeRates() method, not to mention that we use arrays of doubles internally. So, all in all, it makes it unnecessarily cumbersome to provide Python and JavaScript bindings for our solvers.

So, we hide as much as possible from our solvers in our API and bindings, but still test things from C++.