KINSOL solver: don't cache instances anymore.
I thought we could safely cache and therefore reuse our KINSOL object, but after trying different things, I am not able to get the same results as when not caching (which should be our reference). To use CVODE+KINSOL to solve a DAE model was never going to be optimal, but at least it would work. So, here we go, we use KINSOL without any caching. At least, it works.