SolverOde: removed the initialise(), reinitialise(), and solve() methods from the API.
We don't want a user to have access to them. The only reason we had them there was so that we could test our solvers, but it really wasn't a good reason, not least since we can test our solver classes by having `#define protected public` and `#define private public` in our tests.