Solver tests: reorganised things.
Indeed, before we were using

#define private public

to get access to the private implementation of a solver and although it was working fine on macOS, it wasn't on Windows and Linux. MSVC and gcc don't allow to redefine `private`, which on hindsight is fair enough. So, now, we only use the API and test our solvers through a SedDocument object.