Compiler: reuse the first part of our compilation code.
No matter whether we are compiling for WASM or not, the first part of the compilation can be the same. Not only does this save us some code, but it also means that we are not relying on Emscripten's virtual file system. And, the generated code is a bit faster (~2-4% faster)!