Deploy our WASM to https://opencor.ws/libopencor/downloads/wasm.
Useful for people who want to dynamically import libOpenCOR (and thus saving a lot of space when deploying their app/package).

Regarding the `webassembly` job, we now run it on `ubunutu-22.04` rather than `macos-14` because the `milanmk/actions-file-deployer` action requires access to `realpath` and the version of `realpath` on macOS is not the same in that it doesn't support the `--canonicalize-missing` option for instance.