JavaScript tests: account for garbage collection.
Indeed, we need to account for the fact that we have no control over when garbage collection kicks in. So, for this, we make sure that the file manager gets reset before each test. We do this for all our test suites that make use of files even though it's not always needed.